+++ components/src/core/InfoListItem/InfoListItem.tsx
@@ -29,6 +29,8 @@ export type InfoListItemProps = Omit<ListItemProps, 'title' | 'divider'> & {
      * Default: false
      */
     chevron?: boolean;
+    /** Color override for the chevron icon */
+    chevronColor?: string;
     /** Custom classes for default style overrides */
     classes?: InfoListItemClasses;
     /** Show a row separator below the row */
@@ -113,6 +115,7 @@ const InfoListItemRender: React.ForwardRefRenderFunction<unknown, InfoListItemPr
         // ignore unused vars so that we can do prop transferring to the root element
         /* eslint-disable @typescript-eslint/no-unused-vars */
         backgroundColor,
+        chevronColor,
         fontColor,
         iconAlign,
         iconColor,
@@ -151,7 +154,7 @@ const InfoListItemRender: React.ForwardRefRenderFunction<unknown, InfoListItemPr
                 <Chevron
                     color={'inherit'}
                     role={'button'}
-                    className={clsx(combine('rightComponent'), defaultClasses.flipIcon)}
+                    className={clsx(combine('chevron'), defaultClasses.flipIcon)}
                 />
             );
         }
@@ -258,6 +261,7 @@ InfoListItem.propTypes = {
         avatar: PropTypes.string,
         icon: PropTypes.string,
         rightComponent: PropTypes.string,
+        chevron: PropTypes.string,
         separator: PropTypes.string,
         subtitle: PropTypes.string,
         title: PropTypes.string,