+++ demos/storybook/stories/welcome.stories.tsx
@@ -86,7 +86,9 @@ stories.add('Brightlayer UI React Components', () => {
                 <div className={classes.icon}>
                     <img src={icon} alt="blui-icon" className={'rotate'} />
                 </div>
-                <Typography variant={'h2'}>Brightlayer UI</Typography>
+                <Typography variant={'h2'}>
+                    Brightlayer <strong>UI</strong>
+                </Typography>
                 <Typography variant={'h4'}>React Component Library</Typography>
                 {packageJSON.version && <Typography variant={'subtitle1'}>v{packageJSON.version}</Typography>}
                 <Typography variant={'subtitle1'} className={classes.description}>